Moved session commands to dedicated cog
