+++ novelsave/client/bots/discord/endpoints/__init__.py
@@ -1,2 +1,3 @@
 from .commands import dm, sources
+from .session import SessionCog
 from .download import Download, DownloadHandler
